ivi-input-controller: ivi-controller: Copy over weston internal functions
Both ivi-controller and ivi-input-controller access functions which are
internal to weston, which means the ABI of these functions may change or
they may even be completely removed. So far, this wayland-ivi-extension
was able to link against libweston and use those functions by sheer luck,
but that may eventually run out. Copy over the internal functions as of
weston commit 2c1ed289 ("build: bump to version 9.0.92 for the beta release")

Signed-off-by: Marek Vasut <marex@denx.de>