ivi-controller: new mechanism for screenshot
From newer weston version, it moved the weston_renderer to
internal, so the screenshot functional isn't compatible with newer
weston.

To supporting the screenshot functional, we should move to use
the weston_screenshooter_shoot, an api support to capture screen.
it requires the ivi controller need to change the ivi-wm protocol,
mechanism interactive between client and server

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>