ivi-controller: Don't wait for new frame if planes are already disabled
While taking screenshot we disable the HW planes and wait for a new frame signal
to capture the screen content. Here new frame signal is needed to force disable
planes which in-turn make sure GPU rendering is used.

But there is no need to perform a repaint if the planes are already disabled.

Signed-off-by: Rajendraprasad K J <KarammelJayakumar.Rajendraprasad@in.bosch.com>