ilmInput: support ilm_getDefaultSeat api to return default seat name
Sometimes the client need to know the default seat is
available or not, or it wants to know which seat is default among
multi seats.

The seat_create_event should add more an argument for identification
of default seat. The ilm_getDefaultSeat is going to query the list
of seats to find the default seat name.

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>