ivi-controller: set surface background color from client side
The weston_surface_set_color should be only called to setup
the color of a surface without a buffer attached (BUFFER_TYPE_SOLID).
ivi-controller uses it to set the color of a background surface
with SHM buffer.

weston 10 commit [1] fails re-draw with assertion, if unexpected
shader properties is set

weston 11 commit [2] removes the weston_surface_set_color api,
then we can't use it any more.

On hmi-controller and desktop-shell, filling the image or
color of the backgound is done by client. Just follow this
way in ivi-controller.

[1] https://gitlab.freedesktop.org/wayland/weston/-/commit/0f52da62
[2] https://gitlab.freedesktop.org/wayland/weston/-/commit/b5605ccd

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>