simple-weston-client: Avoid wayland display roundtrip on termination
wl display roundtrip after destroying the stream objects is done (similar to
clients/weston-debug.c) to wait for the server side to close all files.

But unlike weston-debug utility, simple-weston-client is terminated only when
Weston is stopped and performing a display roundtrip while Weston is
terminating is seen to result in display fatal error, as the socket is closed
by compositor.

Use wl display flush instead of roundtrip to resolve the issue.

Signed-off-by: Rajendraprasad K J <KarammelJayakumar.Rajendraprasad@in.bosch.com>